requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.6.0+cu126
torchvision==0.21.0+cu126

einops==0.8.0
einx==0.3.0
huggingface-hub==0.34.4
hydra-core==1.3.2
numpy==1.23.0
omegaconf==2.3.0
PyYAML==6.0.2
scipy==1.11.1
tensorboard==2.20.0
tqdm==4.67.1
diffusers==0.35.1
matplotlib==3.8.0
setuptools==68.2.2
clip @ git+https://github.com/openai/CLIP.git
transformers
pandas
ipykernel